fix(android): parse the catalog version as a long, not an int (#1877)
A catalog version is a 14-digit `YYYYMMDDHHmmss` stamp — about 20 trillion,
four orders of magnitude past `Integer.MAX_VALUE`. `Integer.parseInt` threw
`NumberFormatException` for every real filename, `parseVersion` returned -1,
and `shouldCopy` took its "asset is not versioned" fallback: copy unless a
usable file of the *bundled asset's own name* is on disk.

That is the predicate #1830 replaced. Since the JS bootstrap prunes every
catalog older than the one it opened — the bundled file, by construction —
the fallback re-copies ~54 MB on the main thread on every cold start, which
is exactly the defect the fix was for. iOS was unaffected: it parses into
`Int?`, which is 64-bit.

Verified by compiling the app module (`:app:compileDebugJavaWithJavac`) and
by checking on a real JVM that `Integer.parseInt("20260419210656")` throws
while `Long.parseLong` returns the value.

Note this file is compiled by nothing in CI — the native lint lane covers
only the two plugin modules, not the app module.

+++ modules/apps/mobile/android/app/src/main/java/studio/jiva/shruti/BundledDatabaseHelper.java
@@ -115,7 +115,7 @@ public class BundledDatabaseHelper {
      */
     private static boolean shouldCopy(String assetName, String[] onDisk, File targetDir) {
         String stem = versionStem(assetName);
-        int assetVersion = parseVersion(assetName, stem);
+        long assetVersion = parseVersion(assetName, stem);
         if (assetVersion < 0) {
             // Asset is not versioned — nothing to compare, fall back to
             // "copy unless a usable file of the same name is already there".
@@ -123,7 +123,7 @@ public class BundledDatabaseHelper {
         }
 
         for (String candidate : onDisk) {
-            int version = parseVersion(candidate, stem);
+            long version = parseVersion(candidate, stem);
             if (version < assetVersion) {
                 continue;
             }
@@ -140,8 +140,15 @@ public class BundledDatabaseHelper {
         return dot > 0 ? fileName.substring(0, dot) : fileName;
     }
 
-    /** Version in {@code <stem>.<digits>.db}, or {@code -1} when it does not match. */
-    private static int parseVersion(String fileName, String stem) {
+    /**
+     * Version in {@code <stem>.<digits>.db}, or {@code -1} when it does not match.
+     *
+     * A catalog version is a 14-digit {@code YYYYMMDDHHmmss} stamp, four orders of
+     * magnitude past {@code Integer.MAX_VALUE} — parsing it as an {@code int} threw
+     * on every real filename, so the caller saw {@code -1} and skipped the whole
+     * version comparison.
+     */
+    private static long parseVersion(String fileName, String stem) {
         String prefix = stem + ".";
         String suffix = ".db";
         if (!fileName.startsWith(prefix) || !fileName.endsWith(suffix)) {
@@ -157,7 +164,7 @@ public class BundledDatabaseHelper {
             }
         }
         try {
-            return Integer.parseInt(digits);
+            return Long.parseLong(digits);
         } catch (NumberFormatException e) {
             return -1;
         }
